Projects are referenced
The reference graph is the following.

RockAPI (WebApi) --+--> DomainModel <-----+
                   |                      |
		   +--> CompositionRoot --+

The CompositionRoot references all the projects containing interfaces
and/or implementations, but the WebApi project.
The WebApi project references the CompositionRoot, since it will have
to trigger the wiring. It references the DomainModel as well in order to
use the domain classes.
Note that the DomainModel project does not reference any project.