FakeInterface wired to its implementation
The composition root wires the FakeInterface service to a (fake)
implementation. Note that, since the implementation was internal, the
containing namespace has been marked with an attribute, as follows.

[assembly: InternalsVisibleTo("CompositionRoot")]
namespace DomainModel.FakeInterfaces
{
    ...
}

This attribute allows the composition root to access the implementation.
Anyway, the implementation remains internal and cannot be accessed from
all the other assemblies.