Test suite is available
Each SOLID project should follow Test Driven Development methodology.
Thus, a test suite has been (a little bit late) added in a separate
project (DomainModel.Test). The test suite can be executed by launching

dotnet test

The test project and the test class contained in it are automatically
discovered and the test cases are executed.

Note that the in order to test implementations provided by internal
classes, these classes must be accessible by the test project by using
the InternalsVisibleTo() namespace annotation (so as we did with the
composition root, which for other reasons, must access the internal
classes).