A concrete query to sum two integers exists
In the DomainModel project a query summing two integers is available. It
is based on the base classes provided by the CQRS library.

The GetIntSumQuery is the DTO feeding the query (DTO is a pattern
available here: https://en.wikipedia.org/wiki/Data_transfer_object).

The GetIntSumQueryHandler is the class performing the action (it sums
two integers).

These two classes belongs in the DomainModel because the perform a
business action.

Note that the DomainModel now references the CQRS library.