A new controller can sum two integers
The IntSumCalculator controlles can sum two integers and exposes such an
action through the Get method. The controller injects the query abel to
sum the integers, in the form of an instance of this type:

IQueryHandler<GetIntSumQuery, int>

As a result, the WebApi project (containing the controller) now
references the CQRS library. The method can be called through a URL such
as this (you can run it with a browser or with a REST client):

http://baseUrl/api/IntSumCalculator&first=4&second=3

So far, the application compiles but it gives a runtime error since the composition
root verification fails with the SimpleInjector Verify() method. The
composition root must be configured in order for this problem to be solved. It
will be done in the next commit.