Composition root configured for CQRS wiring
Composition root is configured to perform batch registration of the generic
classes performing the CQRS actions. The batch registration uses the great
generics support provided by SimpleInjector. The CompositionRoot project
now references the CQRS project.

The action method can be executed and it provides the expected results.
A sample URL to be called is the following:

http://baseUrl/api/IntSumCalculator?first=4&Second=14