Log project is available
A new project in charge of configuring and providing a ILogger instance
is available in the solution. The project is based on the Serilog
library.

So far, the project is configured to log starting from the Debug
level (the serilog.sinks.trace library is installed) and logs on the
trace window (available during project run in the Output window
selecting the Debug output).

The Startup.cs file contains the command to configure the logger. Thus,
the WebApi project references the logger project.