WRONG way of logging enabled
Systematic query and command logging enables auditing and this is
highly desirable.

In this version GetIntSum query is logged. Two log lines are generated
for each execution.

The first:

2019-04-27 16:38:17.697 +02:00 [Information] Action starting "DomainModel.CQRS.Queries.GetIntSum.GetIntSumQuery": "{\"First\":4,\"Second\":13}"

This line is generated BEFORE the action is executed and it reports the
execution time, the class of the query being executed, the DTO value
(who executes the query and the id of the executing thread are missing
so far, but we will fix this later).

The second:

2019-04-27 16:38:17.702 +02:00 [Information] Action executed (0 ms)

This line is generated AFTER the action is executed and it just reports
the execution time.

Nevertheless, this approach is wrong. Logging is a cross-cutting concern,
but in this approach it is not treated as such: each query must add
this (ugly) logging code and, in case we forget to do it, the log would
be missing with no evidence.

So, the generated log is useful and correct, but the approach is wrong.
In the following commits we will revert to the previous version and we
will follow a more correct, aspect-oriented, approach.