A decorator provides the log cross-cutting concern
In the Logging project, the new class QueryHandlerLogDecorator provides
logging functionality according to an Aspect Oriented approach.

This class is a decorator which wraps the query handler (the decorated
instance) and logs messages both BEFORE and AFTER the action is actually
executed. The logged messages are exactly those logged two commits ago,
but now the log can be set as infrastructural through dependency
injection. This will happen in the following commit.

Note that, so far, commands are still not handled. This will be fixed in an
upcoming commit.