AddUserFake is an available command
A new command has been created: AddUserFake. This command fakes the
insertion of a new user, and its implementation is empty.

A new test case is available to test that a user can be added.

So far, the command is still not exposed through a controller; nor
logging is configured to be executed whenever a commands is fired.
This will be done in the next commits.