Classes aimed at validating commands DTOs
The CQRS project contains the classes used to validate commands.
Validators inherit from ICommandHandler<> and their Handle() method is
in charge of validating the DTO the command is called with. In case
validation is unsuccessful, the validator returns one or more validation
results. The validation is actually performed by a
ValidatedCommandHandlerDecorator which, as its name suggests, is a
command decorator. As the other decorators, it can be systematically
wired every time a command is instantiated, and the container is able to
associate the correct validator (if any) to each command. Again, the
association is based on the concrete type of the DTO. The decorator
throws an execption in case of unsuccessful validation; otherwise it
executes the action of the (decorated) command.

The composition root contains the batch-registering rules to correctly
compose the decorator chain.

Many thanks to Steven van Deursen for having suggested this approach,
here: https://stackoverflow.com/questions/52593426/registering-open-generic-decorators-through-a-single-binding-rule