A validator is available for AddUserFake command
Now that the validatong decorator chain is configured and validators are
automatically discovered by the DI container, it is time to wire a
validator. It is just a matter of creating a validator class which uses
the same DTO of the command to be validated.

As you can see, just creating this file is enough: the decorator
correctly validates the command DTO. You can try it by firing an action
with an empty username and/or role.