Validation is provided for queries as well
So as for the commands, validation infrastructure is provided also for
the queries. Composition root handles the query validation injection.

Note that the classes ValidationException and ValidationResult have been
moved within a new folder, one level up. In fact, these classesi are
used by both the command validators and the query validators.