Sample query validation added
Now that the validation infrastructure is integrated in the framework,
doing a validation on a query is a matter o creating a plain class
inheriting from IQueryValidator<TInDto, TOutDto> (exactly as it was for the
commands).

According to this principle, a class is available as a validator:

  public class OnlyPositiveNumbersValidator : IQueryValidator<GetIntSumQuery, int>

which checks whether both First and Second DTO attributes are greater
than zero. It is easy to check that if you execute the action with a
non-positive number, a validation exception is thrown.