Composition root configured for authorizers
Authorizers are wired as decorators by the composition root. Note that
the order of binding rules is important since the decorators form a
chain and act in the order they are chained. The rules bind the Command
first. Then they are decorated with the authorizers. Then they are
decorated with the validators, then with the logger. This way, the chain
appears as follows:

Logger --> Validator --> Authorizer --> Command

and this chain is the right way to order the different responsibilities.