Sample authorization class for a command
A sample class verifies whether the user that is going to be added is
the admin user. In case it is, the authorization fails.

Enabling this authorization is a matter of adding a simple class to the
project. It is automatically wired as a decorator within che command
execution chain.