Authorization infrastructure available for queries
As for the commands, the Authorization cross-cutting concern is supported by two new classes for the query hierarchy, as well.

In the next commits we are going to enable authorizers as decorators within the composition root, and to create a sample query authorizer to show how it is simple to perform an authorization by simply creating a plain class.