Notification infrastructure available for commands
Notification on action execution is another common cross-cutting concern and is in charge of possibly notifiying other systems about the change in the state of this system. Thus, we integrated notification in the same way we did for validation and authorization.

Note that, in case you need multiple notifications, you can write many notification classes: all will be automatically chained together and executed just after the action execution.