Sample notification class for a command
A sample class notifies about the user being added. This fake implementation is empty, but it demonstrates that notification are correctly chained through the generated log, which appears as follows:

2019-05-05 16:46:14.772 +02:00 [Information] Action starting "DomainModel.CQRS.Commands.AddUserFake.AddUserFakeCommand": "{\"Username\":\"foo\",\"Role\":\"user\"}"
2019-05-05 16:46:14.788 +02:00 [Debug] Now validating
2019-05-05 16:46:14.816 +02:00 [Debug] Now authorizing
2019-05-05 16:46:14.827 +02:00 [Debug] Now notifying
2019-05-05 16:46:14.832 +02:00 [Information] Action executed (45 ms)

Enabling this notification is a matter of adding a simple class to the project. It is automatically wired as a decorator within che command execution chain. In case more classes are written, all of them are added to the execution chain.