Implementation solution folder is available
The Implementation solution folder is aimed at containing all the projects that offer implementation to the service interfaces defined within the `DomainModel` project `Services` folder. As a good practice, all the classes defined in such a project should be internal. The projects should use the `InternalsVisibleTo()` compiler directive in order to let the implementation classes be referenced by the CompositionRoot project. The correct directive appears as follows.

```
[assembly:InternalsVisibleTo("CompositionRoot")]
```
This way, the only project which can reference implementation classes is the composition root. It can bind implementations to abstract interfaces belonging to the domain model. The other projects cannot see any of those implementation classes (according to the _program to an interface, not an implementation_ principle).